Update completed phase toggle behavior
This commit changes the behavior such that multiple phases can be
toggled to be completed by clicking a later phase. Doing so will toggle
all phases prior to be completed as well. Similarly, de-toggling a
completed phase will also toggle off all phases which occur after it.
This change in behavior means that completed phases are always a
contiguous list starting with the first toggleable phase, followed by a
contiguous list of incomplete phases.